pyproject.toml
[project]
name = "watchout-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "google-genai>=1.46.0",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
